+++ src/OrchardCore.Modules/OrchardCore.Workflows/Http/Filters/WorkflowActionFilter.cs
@@ -44,7 +44,7 @@ namespace OrchardCore.Workflows.Http.Filters
 
             if (workflowTypeEntries.Any())
             {
-                var workflowTypeIds = workflowTypeEntries.Select(x => Int32.Parse(x.WorkflowId)).ToList();
+                var workflowTypeIds = workflowTypeEntries.Select(x => Int64.Parse(x.WorkflowId)).ToList();
                 var workflowTypes = (await _workflowTypeStore.GetAsync(workflowTypeIds)).ToDictionary(x => x.Id);
                 var correlationId = routeValues.GetValue<string>("correlationid");
