+++ src/OrchardCore/OrchardCore.Search.Elasticsearch.Core/Services/ElasticIndexingService.cs
@@ -55,7 +55,7 @@ namespace OrchardCore.Search.Elasticsearch.Core.Services
         public async Task ProcessContentItemsAsync(string indexName = default)
         {
             var allIndices = new Dictionary<string, long>();
-            var lastTaskId = Int32.MaxValue;
+            var lastTaskId = Int64.MaxValue;
             IEnumerable<ElasticIndexSettings> indexSettingsList = null;
 
             if (String.IsNullOrEmpty(indexName))
