+++ src/OrchardCore/OrchardCore.Workflows.Abstractions/Services/IWorkflowTypeStore.cs
@@ -6,9 +6,9 @@ namespace OrchardCore.Workflows.Services
 {
     public interface IWorkflowTypeStore
     {
-        Task<WorkflowType> GetAsync(int id);
+        Task<WorkflowType> GetAsync(long id);
         Task<WorkflowType> GetAsync(string uid);
-        Task<IEnumerable<WorkflowType>> GetAsync(IEnumerable<int> ids);
+        Task<IEnumerable<WorkflowType>> GetAsync(IEnumerable<long> ids);
         Task<IEnumerable<WorkflowType>> ListAsync();
         Task<IEnumerable<WorkflowType>> GetByStartActivityAsync(string activityName);
         Task SaveAsync(WorkflowType workflowType);
